chainlit-backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
argilla==1.25.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.401
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
comm==0.2.2
cryptography==42.0.5
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.108.0
fastapi-socketio==0.0.10
fastjsonschema==2.19.1
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.22
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.27
Lazify==0.4.0
literalai==0.0.300
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.23.5
olefile==0.47
openai==1.14.1
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20231228
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pypandoc==1.13
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.32.0.post1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.3.0
tiktoken==0.6.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
unstructured==0.7.12
uptrace==1.22.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.18.1
